driver and check
Edited driver_forcing1 to enable it to run again and only make the
forcing if needed.  Also got rid of error messages related to existing
directories.

Some edits to check_all_results to make it better at counting things.

+++ driver/driver_forcing1.sh
@@ -5,6 +5,8 @@
 # allowing for either a forecast or backfill.
 
 # This version (8/21/2016) is designed to handle jobs more gracefully.
+# And on 5/20/2017 I made it so that it could be run again and only re-make
+# the forcing directory if needed. (Need an override flag?)
 
 # NOTE: must be run from fjord.
 
@@ -101,45 +103,61 @@ do
   echo $(date)
   LOogf_fi=$LOogf_f"/Info"
   LOogf_fd=$LOogf_f"/Data"
-  # make sure directories exist
-  mkdir $LOo
-  mkdir $LOog
-  mkdir $LOogf
-  if [ -d $LOogf_f ]
-  then
-    rm -rf $LOogf_f
-  fi
-  mkdir $LOogf_f
-  mkdir $LOogf_fi
-  mkdir $LOogf_fd
   
-  # Make the forcing.
-  cd $LO_parent"/forcing/"$frc
-  source $HOME"/.bashrc"
-  if [ -e $HOME"/.bash_profile" ] ; then
-    source $HOME"/.bash_profile"
-  fi
-  if [ -e $HOME"/.profile" ] ; then
-    source $HOME"/.profile"
-  fi
-  python ./make_forcing_main.py -g $gridname -t $tag -f $frc -r $run_type -d $DD -x $ex_name > $LOogf_fi"/screen_out.txt" &
-
-  # Check that the job has finished successfully.
-  PID1=$!
-  wait $PID1
-  echo "job completed for" $f_string
-  echo $(date)
+  # this makes all parent directories if needed
+  mkdir -p $LOogf
   
   checkfile=$LOogf_fi"/process_status.csv"
-  # check the checkfile to see if we should continue
-  if grep -q "result,success" $checkfile ; then
-    echo "- job completed successfully."
-    # will continue because we don't change while_flag
-  else
-    echo "- Something else happened."
-    while_flag=1
-    # stop the driver
+  
+  already_done_flag=0
+  # check to see if the job has already completed successfully
+  if [ -f $checkfile ]; then
+    if grep -q "result,success" $checkfile ; then
+      echo "- No action needed: job completed successfully already."
+      already_done_flag=1
+    else
+      echo "- Trying job again."
+    fi
   fi
+  
+  if [ $already_done_flag -eq 0 ]; then
+    
+    if [ -d $LOogf_f ]
+    then
+      rm -rf $LOogf_f
+    fi
+    mkdir $LOogf_f
+    mkdir $LOogf_fi
+    mkdir $LOogf_fd
+    
+    # Make the forcing.
+    cd $LO_parent"/forcing/"$frc
+    source $HOME"/.bashrc"
+    if [ -e $HOME"/.bash_profile" ] ; then
+      source $HOME"/.bash_profile"
+    fi
+    if [ -e $HOME"/.profile" ] ; then
+      source $HOME"/.profile"
+    fi
+    python ./make_forcing_main.py -g $gridname -t $tag -f $frc -r $run_type -d $DD -x $ex_name > $LOogf_fi"/screen_out.txt" &
+
+    # Check that the job has finished successfully.
+    PID1=$!
+    wait $PID1
+    echo "job completed for" $f_string
+    echo $(date)
+  
+    # check the checkfile to see if we should continue
+    if grep -q "result,success" $checkfile ; then
+      echo "- Job completed successfully."
+      # will continue because we don't change while_flag
+    else
+      echo "- Something else happened."
+      while_flag=1
+      # stop the driver
+    fi
+  
+  fi # end of already_done_flag test
   
   # This function increments the day.
   # NOTE: it changes y, m, d, and D, even in the scope of this shell script!
